requirements.txt
APScheduler==3.6.3
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
Flask==1.0.2
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
pluggy==0.11.0
py==1.8.0
pytest==4.5.0
python-telegram-bot==13.5
pytz==2021.1
six==1.12.0
tornado==6.1
tzlocal==2.1
wcwidth==0.1.7
Werkzeug==0.15.5
